Fix : 마이페이지 이번달 기록 조회

+++ greenmap/src/main/java/com/ucamp/greenmap/member/dto/response/RecodeResponse.java
@@ -9,12 +9,12 @@ import lombok.Setter;
 @Getter
 public class RecodeResponse {
     private Long memberId;
-    private Integer pointSum;
-    private  Integer verifyTimes; // 인증 횟수
+    private Long pointSum;
+    private  Long verifyTimes; // 인증 횟수
     private String mostKind;// 가장 많이 한 활동 네임
-    private Integer mostTimes;// 가장 많이 한 활동 횟수
-    private Integer timesDiff;// 지난달 대비 인증 횟수
-    private Integer pointDiff; // 지난달 대비 획득 포인트
+    private Long mostTimes;// 가장 많이 한 활동 횟수
+    private Long timesDiff;// 지난달 대비 인증 횟수
+    private Long pointDiff; // 지난달 대비 획득 포인트
